refactor - #47 IdConverter util에서 auth로 변경
