+++ src/main/java/sopt/org/HMH/domain/challenge/controller/ChallengeController.java
@@ -13,8 +13,8 @@ import sopt.org.HMH.domain.challenge.dto.request.ChallengeRequest;
 import sopt.org.HMH.domain.challenge.dto.response.AddChallengeResponse;
 import sopt.org.HMH.domain.challenge.dto.response.ChallengeResponse;
 import sopt.org.HMH.domain.challenge.service.ChallengeService;
+import sopt.org.HMH.global.auth.UserId;
 import sopt.org.HMH.global.common.response.ApiResponse;
-import sopt.org.HMH.global.util.IdConverter;
 
 import java.security.Principal;
 
@@ -31,7 +31,7 @@ public class ChallengeController {
         return ResponseEntity
                 .status(ChallengeSuccess.ADD_CHALLENGE_SUCCESS.getHttpStatus())
                 .body(ApiResponse.success(ChallengeSuccess.ADD_CHALLENGE_SUCCESS,
-                        challengeService.addChallenge(IdConverter.getUserId(principal), request.period(), request.goalTime())));
+                        challengeService.addChallenge(UserId.getUserId(principal), request.period(), request.goalTime())));
     }
 
     @GetMapping
@@ -40,6 +40,6 @@ public class ChallengeController {
         return ResponseEntity
                 .status(ChallengeSuccess.GET_CHALLENGE_SUCCESS.getHttpStatus())
                 .body(ApiResponse.success(ChallengeSuccess.GET_CHALLENGE_SUCCESS,
-                        challengeService.getChallenge(IdConverter.getUserId(principal), os)));
+                        challengeService.getChallenge(UserId.getUserId(principal), os)));
     }
 }
