+++ src/main/java/sopt/org/HMH/global/auth/UserId.java
@@ -1,14 +1,12 @@
-package sopt.org.HMH.global.util;
-
-import sopt.org.HMH.global.auth.jwt.exception.JwtError;
-import sopt.org.HMH.global.auth.jwt.exception.JwtException;
-
-import java.security.Principal;
+package sopt.org.HMH.global.auth;
 
 import static java.util.Objects.isNull;
 
-public class IdConverter {
+import java.security.Principal;
+import sopt.org.HMH.global.auth.jwt.exception.JwtError;
+import sopt.org.HMH.global.auth.jwt.exception.JwtException;
 
+public class UserId {
     public static Long getUserId(Principal principal) {
         if (isNull(principal)) {
             throw new JwtException(JwtError.EMPTY_PRINCIPLE_EXCEPTION);
